Visnum/labo1/pyproject.toml
[tool.poetry]
name = "labo1"
version = "0.1.0"
description = ""
authors = ["thomas.dagierjo <thomas.dagier-joncour@etu.hesge.ch>"]

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "^3.3.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
Pillow = "^8.1.0"
numpy = "^1.20.1"
matplotlib = "^3.3.4"
black = "^20.8b1"
mypy = "^0.812"
typing = "^3.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
